docs: fix 20+ inaccuracies found by deep cross-reference audit
Systematic audit of all 34 doc files against the Go source code,
schema SQL, and proto definitions. Fixes factual errors only.

High severity:
- grpc-api.md: expires_at "Unix seconds" → "epoch nanoseconds"
- integrate-clp.md: clp-s command --single-file-archive → -o flag
- configure-tables.md: remove 2 unimplemented feature flags
- configuration.md: DefaultQueryLimit mislabeled as page size
- grpc-api.md: add 3 missing filterable columns

Medium severity:
- scale-workers.md: backoff range 1s-32s → 2s-30s
- early-termination.md: INT UNSIGNED → BIGINT throughout
- 4 files: DynamicIndexManager → IndexManager
- 2 files: IngestBatch → Ingest RPC name
- metadata-tables.md: ext BLOB → MEDIUMBLOB
- troubleshooting.md: fake error format → actual zap log format

Low severity:
- quickstart.md: replace 8 fictitious env vars with 2 real ones
- schema.sql: sketch comment s01..s32 → s01..s64
- metastore/model.go: "Two lifecycle chains" → "Three"
- coordinator-ha.md: DDL column order to match schema.sql
- consolidation.md: generic policy names → actual spark_job/
  time_window/audit names from code

Code fixes found during audit:
- coordinator_unit.go: partition lookahead 3 → 7 days
- taskqueue/model.go: timestamp fields uint32 → int64

+++ internal/taskqueue/model.go
@@ -22,9 +22,9 @@ type Task struct {
 	TableName   string
 	State       TaskState
 	WorkerID    sql.NullString
-	CreatedAt   uint32
-	ClaimedAt   sql.NullInt32
-	CompletedAt sql.NullInt32
+	CreatedAt   int64
+	ClaimedAt   sql.NullInt64
+	CompletedAt sql.NullInt64
 	RetryCount  uint8
 	Input       []byte
 	Output      []byte
